Split to more focused utility libraries
Move some utility functions out of `test_api/src/utils.dart` into more
focused libraries. Where possible move utilities into `test_core` or
`test` if they are not used `test_api`.

Make two copies of the `indent` utility since the usages in `test_api`
and `test_core` have non-overlapping sets of arguments used and it's
simpler to implement each case individually. Refactor some calls to
`prefixLines` into calls to `indent`.

Make a private copy of `unawaited` in the one library that needs it in
`test_api`. Add a copy in `test_core` for the remaining uses. Use
`pedantic` in tests since it is available as a dev dependency.
