+++ pkgs/test/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.16.9-dev
+
 ## 1.16.8
 
 * Fix an issue where coverage collection could hang on Chrome.
