+++ pkgs/test/lib/src/runner/browser/browser.dart
@@ -7,9 +7,10 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:typed_data/typed_data.dart';
+import 'package:pedantic/pedantic.dart';
 
-import 'package:test_api/src/utils.dart'; // ignore: implementation_imports
 import 'package:test_core/src/runner/application_exception.dart'; // ignore: implementation_imports
+import 'package:test_core/src/util/dart.dart'; // ignore: implementation_imports
 
 /// An interface for running browser instances.
 ///
