+++ pkgs/test/pubspec.yaml
@@ -1,5 +1,5 @@
 name: test
-version: 1.16.8
+version: 1.16.9-dev
 description: A full featured library for writing and running Dart tests.
 repository: https://github.com/dart-lang/test/blob/master/pkgs/test
 
@@ -32,8 +32,8 @@ dependencies:
   webkit_inspection_protocol: ^1.0.0
   yaml: ^3.0.0
   # Use an exact version until the test_api and test_core package are stable.
-  test_api: 0.3.0
-  test_core: 0.3.19
+  test_api: 0.3.1
+  test_core: 0.3.20
 
 dev_dependencies:
   fake_async: ^1.0.0
