+++ pkgs/test_api/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.3.1-dev
+
 ## 0.3.0
 
 * **Breaking** `TestException.message` is now nullable.
