+++ pkgs/test_api/lib/src/frontend/prints_matcher.dart
@@ -6,7 +6,7 @@ import 'dart:async';
 
 import 'package:matcher/matcher.dart';
 
-import '../utils.dart';
+import '../util/pretty_print.dart';
 import 'async_matcher.dart';
 import 'expect.dart';
