+++ pkgs/test_api/pubspec.yaml
@@ -1,5 +1,5 @@
 name: test_api
-version: 0.3.0
+version: 0.3.1-dev
 description: A library for writing Dart tests.
 homepage: https://github.com/dart-lang/test/blob/master/pkgs/test_api
 
@@ -11,7 +11,6 @@ dependencies:
   boolean_selector: ^2.1.0
   collection: ^1.15.0
   meta: ^1.3.0
-  path: ^1.8.0
   source_span: ^1.8.0
   stack_trace: ^1.10.0
   stream_channel: ^2.1.0
