+++ pkgs/test_api/test/frontend/add_tear_down_test.dart
@@ -5,7 +5,7 @@
 import 'dart:async';
 
 import 'package:async/async.dart';
-import 'package:test_api/src/utils.dart';
+import 'package:pedantic/pedantic.dart';
 import 'package:test/test.dart';
 
 import '../utils.dart';
