+++ pkgs/test_api/test/frontend/matcher/throws_test.dart
@@ -4,7 +4,7 @@
 
 import 'dart:async';
 
-import 'package:test_api/src/utils.dart';
+import 'package:pedantic/pedantic.dart';
 import 'package:test/test.dart';
 
 import '../../utils.dart';
