+++ pkgs/test_core/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.3.20-dev
+
 ## 0.3.19
 
 * Disable stack trace chaining by default.
