+++ pkgs/test_core/lib/src/runner/configuration.dart
@@ -13,17 +13,17 @@ import 'package:source_span/source_span.dart';
 import 'package:test_api/src/backend/platform_selector.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/runtime.dart'; // ignore: implementation_imports
 import 'package:test_api/src/frontend/timeout.dart'; // ignore: implementation_imports
-import 'package:test_api/src/utils.dart'; // ignore: implementation_imports
 
 import '../util/io.dart';
-import 'suite.dart';
-import 'runtime_selection.dart';
 import 'configuration/args.dart' as args;
 import 'configuration/custom_runtime.dart';
 import 'configuration/load.dart';
 import 'configuration/reporters.dart';
 import 'configuration/runtime_settings.dart';
+import 'configuration/utils.dart';
 import 'configuration/values.dart';
+import 'runtime_selection.dart';
+import 'suite.dart';
 
 /// The key used to look up [Configuration.current] in a zone.
 final _currentKey = Object();
