+++ pkgs/test_core/lib/src/runner/parse_metadata.dart
@@ -6,12 +6,13 @@ import 'package:analyzer/dart/analysis/utilities.dart';
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:path/path.dart' as p;
 import 'package:source_span/source_span.dart';
-
 import 'package:test_api/src/backend/metadata.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/platform_selector.dart'; // ignore: implementation_imports
 import 'package:test_api/src/frontend/timeout.dart'; // ignore: implementation_imports
 import 'package:test_api/src/utils.dart'; // ignore: implementation_imports
+
 import '../util/dart.dart';
+import '../util/pair.dart';
 
 /// Parse the test metadata for the test file at [path] with [contents].
 ///
