+++ pkgs/test_core/lib/test_core.dart
@@ -15,13 +15,14 @@ import 'package:test_api/src/backend/declarer.dart'; // ignore: implementation_i
 import 'package:test_api/src/backend/invoker.dart'; // ignore: implementation_imports
 import 'package:test_api/src/frontend/timeout.dart'; // ignore: implementation_imports
 import 'package:test_api/src/frontend/utils.dart'; // ignore: implementation_imports
-import 'package:test_api/src/utils.dart'; // ignore: implementation_imports
 
 import 'src/runner/engine.dart';
 import 'src/runner/plugin/environment.dart';
 import 'src/runner/reporter/expanded.dart';
 import 'src/runner/runner_suite.dart';
 import 'src/runner/suite.dart';
+import 'src/util/async.dart';
+import 'src/util/io.dart';
 import 'src/util/print_sink.dart';
 
 export 'package:matcher/matcher.dart';
